[compression] Add cache compression on process exit.
Compress largest files first, and only compress enough files
to stay under the maxCacheSize argument.

TODO -
 - Delete oldest files in cache if not enough can be compressed.
 - Iterate on compression until under the limit.

+++ __tests__/index.test.js
@@ -9,7 +9,14 @@ const lowIceChest = require('../lib/low-level');
 const cacheLocation = path.resolve(path.join(__dirname, '.cache'));
 
 beforeEach(() => {
+  if (process.listeners('beforeEach').length > 0) {
+    throw new Error('Whoa! Some assumptions are broken, check me out!');
+  }
   rimraf.sync(cacheLocation);
+});
+
+afterEach(() => {
+  process.removeAllListeners('beforeExit');
 });
 
 test('creating and instance of IceChest should create the cacheLocation', () => {
@@ -29,4 +36,23 @@ test('iceChest instances wrap low-level functions', () => {
   expect(lowIceChest.writeToCache).toBeCalledWith(fakeCacheKey, obj, cacheLocation);
   iceChest.loadFromCache(fakeCacheKey);
   expect(lowIceChest.loadFromCache).toBeCalledWith(fakeCacheKey, cacheLocation);
+});
+
+test('iceChest hooks up a beforeExit hook when passing a maxCacheSize', () => {
+  expect(process.listeners('beforeExit').length).toBe(0); // beforeExit not hooked up.
+  const iceChest = new IceChest({
+    cacheLocation,
+    maxCacheSize: 75, // A single file should exceed the cache size.
+  });
+
+  expect(process.listeners('beforeExit').length).toBe(1);
+  expect(process.listeners('beforeExit')[0]());
+  // expect that performCacheMaintenance was called w/ correct args.
+  expect(lowIceChest.performCacheMaintenance).toBeCalledWith(75, cacheLocation);
+});
+
+test('iceChest does not hook up beforeExit call when not maxCacheSize is passed', () => {
+    expect(process.listeners('beforeExit').length).toBe(0);
+    const iceChest = new IceChest({ cacheLocation });
+    expect(process.listeners('beforeExit').length).toBe(0);
 });
